+++ CHANGES.rst
@@ -11,6 +11,11 @@ unreleased
   not sortable on Python 3 due to differing types.
   See https://github.com/Pylons/pyramid/pull/3457
 
+- Avoid configuring logging in the monitor process usin the logging config
+  intended for the application. This avoids opening files for writing in both
+  processes which can cause issues on some systems.
+  See https://github.com/Pylons/pyramid/pull/3460
+
 .. _changes_1.10.1:
 
 1.10.1 (2018-11-06)